// frontend/src/app/api/schema.d.ts
export interface components {
    schemas: {
// ...
        BrandDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            description?: string | null;
            logoUrl?: string | null;
            productsId?: number[] | null;
        };
// ...
        CategoryDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            description?: string | null;
            imageUrl?: string | null;
            /** Format: int32 */
            parentCategoryId?: number | null;
            type?: string | null;
            productsId?: number[] | null;
        };
        /**
         * Format: int32
         * @enum {integer}
         */
        CategoryType: 0 | 1;
        Color: {
            /** Format: int32 */
            id?: number;
            name: string | null;
            hexCode?: string | null;
            products?: components["schemas"]["Products"][] | null;
        };
        ColorDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            hexCode?: string | null;
        };
        CreateBrandDto: {
            name: string | null;
            description?: string | null;
            logoUrl?: string | null;
        };
        CreateCategoryDto: {
            name?: string | null;
            description?: string | null;
            imageUrl?: string | null;
            /** Format: int32 */
            parentCategoryId?: number | null;
            type?: string | null;
            productsId?: number[] | null;
        };
        CreateColorDto: {
            name?: string | null;
            hexCode?: string | null;
        };
        CreateDiscountDto: {
            /** Format: int32 */
            productId?: number;
            /** Format: double */
            percentage?: number;
            /** Format: date-time */
            startDate?: string;
            /** Format: date-time */
            endDate?: string;
        };
        CreateOrderDto: {
            items?: components["schemas"]["CreateOrderItemDto"][] | null;
        };
        CreateOrderItemDto: {
            /** Format: int32 */
            productId?: number;
            /** Format: int32 */
            quantity?: number;
        };
        CreateProductDto: {
            name?: string | null;
            /** Format: int32 */
            brandId?: number;
            description?: string | null;
            details?: string | null;
            /** Format: double */
            price?: number;
            productCode?: string | null;
            /** Format: int32 */
            stockQuantity?: number;
            status?: string | null;
            /** Format: int32 */
            colorId?: number;
            /** Format: int32 */
            sizeId?: number;
            composition?: string | null;
            categoriesId?: number[] | null;
            matchProductsId?: number[] | null;
            imageUrl?: string[] | null;
        };
        CreateSizeDto: {
            name?: string | null;
        };
// ...
        DiscountDto: {
            /** Format: int32 */
            id?: number;
            /** Format: double */
            percentage?: number;
            /** Format: date-time */
            startDate?: string;
            /** Format: date-time */
            endDate?: string;
        };
        LoginRequestDto: {
            login?: string | null;
            password?: string | null;
        };
        OrderDto: {
            /** Format: int32 */
            id?: number;
            userName?: string | null;
            /** Format: double */
            amount?: number;
            paymentStatus?: string | null;
            status?: string | null;
            /** Format: date-time */
            createdAt?: string;
            items?: components["schemas"]["OrderItemDto"][] | null;
        };
        OrderItemDto: {
            /** Format: int32 */
            productId?: number;
            productName?: string | null;
            /** Format: int32 */
            quantity?: number;
            /** Format: double */
            unitPrice?: number;
        };
        ProductDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
            /** Format: int32 */
            brandId?: number;
            description?: string | null;
            details?: string | null;
            /** Format: double */
            price?: number;
            productCode?: string | null;
            /** Format: int32 */
            stockQuantity?: number;
            status?: string | null;
            /** Format: int32 */
            colorId?: number;
            colorName?: string | null;
            /** Format: int32 */
            sizeId?: number;
            sizeName?: string | null;
            composition?: string | null;
            categoriesId?: number[] | null;
            matchProductsId?: number[] | null;
            imageUrl?: string[] | null;
            discounts?: components["schemas"]["DiscountDto"][] | null;
        };
// ...
        SizeDto: {
            /** Format: int32 */
            id?: number;
            name?: string | null;
        };
        UpdateBrandDto: {
            name: string | null;
            description?: string | null;
            logoUrl?: string | null;
        };
        UpdateCategoryDto: {
            name?: string | null;
            description?: string | null;
            productsId?: number[] | null;
        };
        UpdateProductDto: {
            name?: string | null;
            /** Format: int32 */
            brandId?: number;
            description?: string | null;
            details?: string | null;
            /** Format: double */
            price?: number;
            productCode?: string | null;
            /** Format: int32 */
            stockQuantity?: number;
            status?: string | null;
            /** Format: int32 */
            colorId?: number;
            /** Format: int32 */
            sizeId?: number;
            composition?: string | null;
            categoriesId?: number[] | null;
            matchProductsId?: number[] | null;
            imageUrl?: string[] | null;
        };
// ...
        UserDto: {
            id?: string | null;
            userName?: string | null;
            email?: string | null;
            phoneNumber?: string | null;
            roles?: string[] | null;
        };
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}